pyproject.toml
[tool.poetry]
package-mode = false
name = "automate-groceries"
version = "0.1.0"
description = ""
authors = ["niklas168 <niklaskoenigsdorff@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
bring-api = "^0.9.1"
python-dotenv = "^1.0.1"
solara = "^1.43.0"
sqlalchemy = "^2.0.37"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
